+++ Elements/test/PlaneTests.cs
@@ -1,5 +1,6 @@
 using Xunit;
 using Elements.Geometry;
+using System;
 
 namespace Elements.Tests
 {
@@ -14,6 +15,29 @@ namespace Elements.Tests
 
             var p2 = new Plane(Vector3.Origin, new Vector3(0.1, 0.1, 0.1).Unitized());
             Assert.False(p.IsCoplanar(p2));
+        }
+
+        [Fact]
+        public void ThreePlanesIntersect()
+        {
+            var origin = new Vector3(2, 2, 2);
+            var a = new Plane(origin, Vector3.XAxis);
+            var b = new Plane(origin, Vector3.ZAxis);
+            var c = new Plane(origin, Vector3.YAxis);
+            var intersects = a.Intersects(b, c, out Vector3 result);
+            Assert.True(intersects);
+            Assert.True(result.IsAlmostEqualTo(origin));
+        }
+
+        [Fact]
+        public void ThreePlanesDoNotIntersect()
+        {
+            var origin = new Vector3(2, 2, 2);
+            var a = new Plane(origin, Vector3.XAxis);
+            var b = new Plane(new Vector3(1, 1, 1), Vector3.XAxis);
+            var c = new Plane(origin, Vector3.YAxis);
+            var intersects = a.Intersects(b, c, out Vector3 result);
+            Assert.False(intersects);
         }
     }
 }